Move utilities into lib/utilities directory
