+++ src/app/containers/RadioSchedule/index.stories.jsx
@@ -7,7 +7,7 @@ import { RequestContextProvider } from '#contexts/RequestContext';
 import { ServiceContextProvider } from '#contexts/ServiceContext';
 import { ToggleContextProvider } from '#contexts/ToggleContext';
 import WithTimeMachine from '#testHelpers/withTimeMachine';
-import { getLocalRadioScheduleEndpoint } from './endpoint';
+import { getLocalRadioScheduleEndpoint } from '#lib/utilities/getRadioSchedulesUrls';
 
 // Currently, only these services have radio schedule data
 const validServices = [
